plugins: new flag to set the need for a flag
As described in lxde/lxqt#636, some plugins implement their own context
menus. That way the default context menu, which has the "Configure",
"Move" and "Remove" buttons, can't be activated.

This patch adds a way for themes to provide a little "handle" object
near the plugins that have the flag set, so that the menu can be
activated.

Fixes lxde/lxqt#636.