mainmenu: Fix close menu by "weird" shortcut
Globalkeys uses not compatible string representation of shortcut (it's probably done by XKeysymToString)
with QKeySequence. So we are not able to know if the shortcut sequence was pressed in event handling.
We use a workaround and closing the menu after any "modifier" key pushed (this way also other QMenu objects
behave)