panel: Correct PanelPluginsModel logic
In Qt Model/View architecture, the model must be view agnostic to be
able to show the model in as many views as it is required.

With tracking of current active index, our PanelPluginsModel couldn't
be used in more than one view. This commit fixes that.

This change also fixes the lxde/lxqt#959, as single/double-click
*activated* was fixed in lxde/liblxqt@83c93de.