# This is a combination of 3 commits. # The first commit's message is: Add a configuration dialog for system tray that enables setting tray icon size
# This is the 2nd commit message:

Tray icon size is now adjusted, but a resize of the icon space is necessary

# This is the 3rd commit message:

We just had to override the stylesheet provided in the panel code. Yet, this is an explanation of why user-defined stylesheets were doing nothing to tray icons or taskbar buttons regarding qproperty-iconSize. I think this beahvior should be redefined.