Fix hiding of panel with drag
If something is dragged into the 4-pix zone of an auto-hiding panel, the panel will be shown but, in case it contains the quick launch widget, if the dragging cursor enters the child widget, `QEvent::DragLeave` will be sent, so that `QEvent::DragLeave` won't be sent again if the cursor leaves the child widget and the panel *simultaneously*. As a result, the panel won't be hidden in such cases (until an ordinary cursor enters and leaves it).

This patch fixes the above-mentioned situation with a new (virtual) method, namely `hideOnCursorLeave()`, which is called by `LXQtQuickLaunch::dragLeaveEvent()`. The function checks if the cursor is outside the interior of the panel (on its boundary) and only then, stops the delayed show timer and hides the panel.

The same method may be needed to applied for other DND-enabled widgets. This is only for quick launcher