panel: Hide correctly upon the DragLeave
The DragLeave event is not delivered to the parent widget while DND and
the cursor leaves the parent and child (which also observes drag
events) simulatenously.

The panel needs to know about the DragLeave event (to correctly hide)
even if there is no spare space between the child plugin and the panel
edge.

So we're observing the DragLeave events for all the plugins and its
children widgets.