Optional overriding of icon theme for panels
Closes https://github.com/lxqt/lxqt/issues/470

When a light widget style is used with a dark panel, the symbolic SVG icons follow the style and become dark on the panel. The right way of solving such issues is assigning a dark widget style to the panel but neither Fusion nor Breeze can do that.

This patch is about the "wrong way". It gives a separate icon set to the panel with unavoidable side effects but is optional and may be used as the last resort.