DirectoryMenu: Added Open with terminal option
Added feature to open folder in a terminal. Default terminal is
QTerminal, but it can be changed in the configure screen.

Terminals are now spawned in their own process. Instead of being
called in a `QProcess` instance that is tied to the panel, terminals
are called in `QProcess::startDetached`. This way, they can outlive
the panel should it be terminated. Also, memory leaks are eliminates
because there are no longer `QProcess` pointers floating around.