Only show group popup on left clicking a grouping task button
And do nothing with its corresponding windows.

IMHO, the left click behavior of the task manager is counter-intuitive when there is a group of two or more buttons because the user cannot predict what will happen to which window of the same app on clicking its single task button (unless he remembers the order of appearance of its windows).

This patch prevents confusion by showing the group popup in such cases, so that the user could choose what to do with each window. That's similar to the behavior of KDE Plasma's task manager too.

Closes https://github.com/lxqt/lxqt/issues/1725