Smoother menu search
While the user is typing in the search field, LXQt Panel does lots of computations (searching for the string that's typed at the moment in lots of desktop entries and updating the menu by considering the found matches), which can result in input blocking and GUI freeze, especially with fast typing. To test that, you could keep the key "A" pressed in the search filed if you have an enough number of menu items in you main menu.

This patch starts searching (filtering) only when the user stops typing. The average typing speed is 200 characters per minute. Here, an interval of 350 ms is used for a smoother filtering.