mainmenu: Filter applications more quickly when searching.
Match only from the start of words.  E.g. if I type "ge", assume I
want "GEany" and not "diffuse merGE tool".  This is similar to the
way that tab-completion works in Bash, and saves typing by reducing
the number of results much more quickly.

This also allows shorthand such as "l c" for "Libreoffice Calc".