Added an option to toggle volume notifications with keyboard
If "Always notify about…" is checked, the new box "Notify about volume changes with keyboard" will be disabled and checked. If the user unchecks the former, the latter will be enabled. Finally, if the user unchecks the second box, all notifications will be disabled.

The change is backward-compatible because the default value of the new key is "true".

Closes https://github.com/lxqt/lxqt-panel/issues/1171