quicklaunch: Remove press-and-hold popup menu from launchers.
It's unnecessary because we already have the right-click context
menu which provides the same options (and then some).

This fixes the issue of the launcher buttons being too wide (#1223)
because Qt no longer allocates space for the little arrow that
indicates the presence of a popup menu.  (The arrow was previously
hidden by overriding paintEvent(), but space was still allocated.)