Also set geometry of panel before showing it
Because some WMs (like openbox) might not position the panel correctly if the geometry isn't set before `QWidget::show()`.

Previously, with `QDesktopWidget`, the geometry was corrected in an unusual way, by connecting to `QDesktopWidget::workAreaResized`. Since that signal was emitted after showing the panel, a flash was sometimes visible with openbox at login.