Fixed high CPU usage of Desktop Switch
Due to a change in KWindowSystem or X11 or…, changing the desktop layout inside `DesktopSwitch::realign()` may call `LXQtPanel::realign`, which calls `DesktopSwitch::realign()` and so, creates an infinite loop. But, actually, the desktop layout needs to be set only when the setting is changed.

The patch is backward compatible.

Fixes https://github.com/lxqt/lxqt-panel/issues/1352