Implemented auto-hiding for Status Notifier
The items can be auto-hidden or always shown/hidden. The default is "Always show".

If there are (auto-)hidden items, an extra button will appear, which shows all items when clicked; then (auto-)hidden items will be made hidden again after 2 seconds if the the mouse cursor leaves Status Notifier.

The auto-hidden items that need attention are shown for 5 minutes (the interval is configurable). Apart from explicit attention requests, an icon change is also interpreted as an attention request with auto-hiding. That's useful, for example, when a battery or WiFi icon changes.

Closes https://github.com/lxqt/lxqt/issues/1012