Hide/show status notifier button more appropriately
Hide the status notifier button if all auto-hiding items have attention and there is no item in the hiding list; show it otherwise.

Previously, the button was shown when auto-hiding items existed, even when all of them had attention and no item was in the hiding list.

Also, optimized the code a little.