Use a stylesheet for progress-bars of Sensors plugin (#1418)
Only a style sheet guarantees that custom colors are applied because not all widget styles use palettes to draw progress bars.

Closes https://github.com/lxqt/lxqt/issues/1810

Explanation:

As Qt doc says (see https://doc.qt.io/qt-5/qwidget.html#palette-prop): "The current style, which is used to render the content of all standard Qt widgets, is free to choose colors and brushes from the widget palette, or in some cases, to ignore the palette (partially, or completely)…"

Therefore, in general, it's a bad practice to use palettes in a code to customize widgets. The correct method is either using Qt style sheets (the easy way) or drawing the customized widgets completely by using `QStyle` (cumbersome).