Implemented auto-hiding for Status Notifier (#1412)
* Implemented auto-hiding for Status Notifier

The items can be auto-hidden or always shown/hidden. The default is "Always show".

If there are (auto-)hidden items, an extra button will appear, which shows all items when clicked; then (auto-)hidden items will be made hidden again after 2 seconds if the the mouse cursor leaves Status Notifier.

The auto-hidden items that need attention are shown for 5 minutes (the interval is configurable). Apart from explicit attention requests, an icon change is also interpreted as an attention request with auto-hiding. That's useful, for example, when a battery or WiFi icon changes.

Closes https://github.com/lxqt/lxqt/issues/1012

* Hide/show status notifier button more appropriately

Hide the status notifier button if all auto-hiding items have attention and there is no item in the hiding list; show it otherwise.

Previously, the button was shown when auto-hiding items existed, even when all of them had attention and no item was in the hiding list.

Also, optimized the code a little.

* Wait for c-tor before announcing the title of a Status Notifier item

Because of a possible race condition.

* Code cleanup for Status Notifier's auto-hiding