Workaround an issue with glibc 2.33 on old Docker engines
Since glibc 2.33, the faccessat() function tries the faccessat2 system
call first and falls back to the faccessat system call on older
kernels [1].  However, the seccomp profile provided by old Docker engines
does not allow faccessat2 (fixed by [2][3]). As a result, the faccessat2
system call returns EPERM, and glibc returns an error in this case. Using
a patched glibc to workaround this issue.

Note that the aforementioned glibc commit does not mean that glibc 2.33
building on new kernels does not work with old kernels. Instead, the
__LINUX_KERNEL_VERSION macro is 0 by default to achieve best-effort
compatibility with various kernel versions.

[1] https://sourceware.org/git/?p=glibc.git;a=patch;h=3d3ab573a5f3071992cbc4f57d50d1d29d55bde2
[2] https://github.com/moby/moby/pull/41353
[3] https://github.com/moby/moby/pull/41381