Prevent flickering of desktop switcher when only active button is shown (#1704)
If only the active button is shown, the layout of desktop switcher might grow and shrink on switching desktop because, in a very short interval, the number of visible buttons will change. That may cause a flickering under rare and random circumstances. When it happens, it can be quite visible because it may shake nearby plugins.

The flickering is prevented here by disabling the layout before changing visibilities of buttons and re-enabling it afterward.