Fixed Quick Launch with multiple config files
The code checks the keys "desktop", "file" and "name/exec/icon" in this order and ignores the rest if one is found. Because of that, for example, if the user config file contains "name/exec/icon" but another config file contains "desktop" *for the same Quick Launch button*, the former will be ignored and the latter will be used instead.

To fix this problem, the patch sets "desktop" and "file" to empty strings for executable scripts. It also sets "desktop" to an empty string for files.

If the user has encountered the bug before this patch, he should either remake his script/file launcher(s) or temporarily change some setting to get rid of the problem.

Fixes https://github.com/lxqt/lxqt-panel/issues/1822