Fixed a regression in volume popup
The code shouldn't try to show or hide the popup based on the enter and leave events of its button because those events aren't (or shouldn't be) emitted when the popup is visible.

Also, the option for showing the popup on button mouseover has been removed, because the popup couldn't be hidden when the cursor left the button.

Previously, the code made use of a Qt issue to do the above-mentioned actions, but the issue was avoided for another purpose in V1.2.0.

Fixes https://github.com/lxqt/lxqt-panel/issues/1839