Refactor LXQtTaskbarWlrootsBackend
Instead of using LXQtTaskbarWlrootsWindow* pointers, it's easier to
store and use WId a.k.a quintptr which are derived from the address of
the LXQtTaskbarWlrootsWindow pointer. This way, it's trivial to retrieve
the underlying pointer with a reinterpret_cast<...>().

Also, switch to using std::find(...) to search a given WId in windows.
Use a dedicated function to erase a WId from windows: eraseWindow(...).