Fixed a crash is worldclock popup on Wayland with Qt 6.8.0
The crash was in `deleteLater()`. The code is simplified, such that there's no need to `deleteLater()`. It's also good for X11.

Fixes https://github.com/lxqt/lxqt-panel/issues/2146

NOTE: This is similar to https://github.com/lxqt/lxqt-notificationd/pull/375 and https://github.com/lxqt/lximage-qt/pull/666, and shows that we should be more cautious when using `deleteLater()`.