Prevent panel from accepting focus with some Wayland compositors
Unfortunately, several Wayland compositors won't let panel popups have focus if the panel does not accept focus.

However, some compositors don't have this issue, and we can use `KeyboardInteractivityNone` with them. The patch does that for `kwin_wayland`, 'wayfire' and 'hyprland', and with Qt ≥ 6.8 (because I'm not sure it works with previous versions of Qt).

NOTE: This is an ugly patch, but the current situation of Wayland compositors isn't better. The patch can be considered as a temporary solution.