Handle Wayland screens by using their names (#2226)
* Handle Wayland screens by using their names

The main reason behind this change is that screen setting by using numbers may not work with all Wayland compositors.

With this patch, and on Wayland,

 * The screen names are used for determining where the panels should be shown;
 * A panel is not moved to another screen when its screen is disconnected (unlike on X11);
 * If the screen assigned to a panel is not connected on startup, that panel won't be created, but it will be created as soon as its corresponding screen is connected.

IMPORTANT NOTE: Since I couldn't test this patch, I made it based on some assumptions. For example, I don't know if Qt or the compositor hides a panel, automatically moves it to another screen, or just causes a crash when its screen is unplugged. Here, I've *assumed* that it hides the panel. So, the patch should be tested thoroughly, and it might need changes.

* Add all panels of an added screen

The user may have assigned multiple panels to that screen.