Enabled PageDown/PageUp for Fancy Menu
Home and End aren't enabled for the views because they can be used in the filter entry (even without a filter, they would be confusing).

NOTE: This is not related to the PR but is worth mentioning: Although the logic of using keyboard is correct, it can still be confusing to the user, especially because there's no visual clue about the focused view. We might want to do something about it later.