Workaround for kwin_wayland's problems in showing desktop
At least there were two problems in KWin's handling of showing desktop: it didn't minimize pcmanfm-qt's windows, and it minimized what it shouldn't, e.g., the windows on the bottom layer or Conky.

Here I circumvented KWin's handling of showing desktop by doing it directly, as I did for the wlroots backend but with small changes. One good thing about KWin is that it doesn't automatically focus a window on unminimizing it. So, the implementation got even better, IMO.

NOTE: kwin_x11 has had exactly the same problems for a long time, but I'm not motivated to log into X11 and make a workaround there — at least not for now ;)