Disabled automatic showing of mount popup on Wayland
In other words, the option "Popup menu" is removed under Wayland. If it's selected on X11 or by editing the config file manually, the selection won't change, but it will be treated as "Show info" and displayed as such by the config dialog.

The reason is an old problem of Qt+Wayland: Before the first input (mouse) interaction with the app, Wayland displays its popup as a standalone window, e.g., decorated under stacking compositors. Worse than that, if the user lets the popup disappear by itself without clicking anywhere (the code has a timer for it), the first click on the mount button will cause a crash (without backtrace).

NOTE:

I've encountered similar Wayland issues elsewhere, e.g., in `QClipboard`. Their common denominator has been QtWayland's dumbness before the first input interaction.

So, this is the rule of thumb: Under Wayland, don't show a popup of a Qt app and don't rely on `QClipboard` *without a input interaction*.