Completed showing current screen windows in wlroots taskbar (#2291)
In addition, the task-bar code had a problem that is fixed here:

When window grouping was enabled and only windows from the current screen/desktop were shown, the checked state of task buttons could be wrong if one window inside a group was sent to another screen/desktop.

I think this issue was very old and existed on X11 too but never reported. Anyway, a thorough test under X11 is really needed.

Closes https://github.com/lxqt/lxqt-panel/issues/2091