ext-workspace-v1 support — cleanup
This is the same as https://github.com/lxqt/lxqt-panel/pull/2323 (@marcusbritanicus' PR) with these changes:

 1. All changes to the existing code style are removed to make the real changes visible.
 2. Showing of `n/a` is fixed (on labwc).
 3. The tooltips are fixed (on labwc).

I suggest we merge this to have not only a minimal switcher support for labwc but also a clean foundation for future changes that @marcusbritanicus might make. Particularly, the first change was necessary for clearly keep track of real code changes and avoiding regressions.

@marcusbritanicus
If my suggestion about merging is not accepted, would you please make your future changes to this branch, instead of yours at https://github.com/lxqt/lxqt-panel/pull/2323? Also, please don't change the code style in other places; we could do that in a separate PR later.

Please note that the behavior of "Show only windows from desktop" has changed. That's inevitable because `LXQtTaskbarWlrootsBackend::getCurrentWorkspace` is working now, while `LXQtTaskbarWlrootsBackend::getWindowWorkspace` is like before. The user should not check that box — disabling it for this backend would *really* complicate the code.